Initial infrastructure for "spectrum mode". Do not upload anywhere!
This is the first commit of the new measurement packet infrastructure
and it needs other parts as well in order to function properly!

Implemented the "spectrum mode" (see Trello for details) with the
theoretical (because I haven't tested it) capability of multipacket
spectrums.
Also started the documentation of the measurement process but its still
in early phase. Expect a commit about that until the final PR.