Fixed the repetition 63 crash (details below!)
Once upon a time we decided to get rid of malloc from the program and
allocate the queue memory from the `queue_init` method using and array.
The problem with this that after the `queue_init` finishes it frees up
its stack memory including the array that we used to init our queue in
RAM. This is why I switched back to `malloc` for this functionality,
because this way the queue memory won't be deallocated when `queue_init`
returns.