Fix for the flash issue. UNSTABLE VERSION!
What was the error?
We experienced that after restart the queues weren't as long as "long"
as they should be. After further checks I realized that the loaded
content is not the same as the one in flash. The problem was that in the
`flash_load` method we used a `uint32_t *` pointer to keep track of the
address and we only read 2 bytes (a `uint16_t`) each time and then we
incremented the address by 2. The problem with this, that if you
increment a `uint32_t *` pointer by 2 it would move 8 bytes, since the
uint32_t is a four byte pointer and we only read 2 of the 8 bytes and we
skipped the other 6. This led to serious malfunctioning of the queue
system and maybe (maybe because I haven't checked yet) this could reason
some of the other wierd things around the Scheduler (unknown states,
wrong timecode, etc...)

How it is fixed?
I modified the the `flash_load` method to read all four bytes of
`uint32_t` and now the address is incremented only by 1. This ensures
that the "loaded content" is exactly the same as the flash version.

Why this version is unstable?
The flash interface is a crucial part of our system and this change
might affected other subsystems that may crash on this state. Stay tuned
until the Pull Request when I've checked every possibility.