Implemented queue cursor saving to Flash.
What is a queue cursor?
I've done a bit of refactoring and as a result now we have a separate
struct for head, tail and size called queue cursor. This was needed
because we wanted to save head, tail and size.
For the saving, loading and handling of these queue cursors, I've
decided to implement a new subsystem called queue manager that is
dedicated for this porpuse. Expect detailed documentation soon.